Moving mrext usage to mister platform
