+++ pkg/platforms/mister/config.go
@@ -1,5 +1,10 @@
 package mister
 
+import (
+	mrextConfig "github.com/wizzomafizzo/mrext/pkg/config"
+	config "github.com/wizzomafizzo/tapto/pkg/config"
+)
+
 const (
 	TempFolder        = "/tmp/tapto"
 	LogFile           = TempFolder + "/tapto.log"
@@ -11,3 +16,20 @@ const (
 	MappingsFile      = "/media/fat/nfc.csv"
 	TokenReadFile     = "/tmp/TOKENREAD"
 )
+
+func UserConfigToMrext(cfg *config.UserConfig) *mrextConfig.UserConfig {
+	return &mrextConfig.UserConfig{
+		AppPath: cfg.AppPath,
+		IniPath: cfg.IniPath,
+		Nfc: mrextConfig.NfcConfig{
+			ConnectionString: cfg.TapTo.ConnectionString,
+			AllowCommands:    cfg.TapTo.AllowCommands,
+			DisableSounds:    cfg.TapTo.DisableSounds,
+			ProbeDevice:      cfg.TapTo.ProbeDevice,
+		},
+		Systems: mrextConfig.SystemsConfig{
+			GamesFolder: cfg.Systems.GamesFolder,
+			SetCore:     cfg.Systems.SetCore,
+		},
+	}
+}
