+++ pkg/platforms/mister/migration.go
@@ -1,4 +1,4 @@
-package utils
+package mister
 
 import (
 	"os"
